+++ Changelog.md
@@ -1,7 +1,8 @@
---------------------------Changelog for FUTURE VERSION of SecurityCraft--------------------------
+--------------------------Changelog for the next version of of SecurityCraft--------------------------
 
 - Fix: Several mod incompatibilities that caused Frame feeds to not show up correctly
 - Fix: Game performance does not improve when breaking frames with an active frame feed
+- Fix: The powered Secure Redstone Interface model looks wrong when placed on the wall or ceiling
 
 --------------------------Changelog for v1.10 of SecurityCraft--------------------------